Adopt DisconnectRequestHandler
